ai-service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
tensorflow-cpu     
requests